init useModel as vue/core currently does it for parity, move to setupHelpers
